authentication for delete and edit added

+++ routes/campgrounds.js
@@ -38,6 +38,7 @@ router.post(
     //   throw new ExpressError("invalid Campground Data", 400);
 
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash("success", "successfully made a new campground");
     res.redirect(`/campgrounds/${campground._id}`);
@@ -47,9 +48,10 @@ router.post(
 router.get(
   "/:id",
   catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate(
-      "reviews"
-    );
+    const campground = await Campground.findById(req.params.id)
+      .populate("reviews")
+      .populate("author");
+    console.log(campground);
     if (!campground) {
       req.flash("error", "Campground not found");
       return res.redirect("/campgrounds");
@@ -62,7 +64,16 @@ router.get(
   "/:id/edit",
   isLoggedIn,
   catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground) {
+      req.flash("error", "Campground not found");
+      return res.redirect(`/campgrounds`);
+    }
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("error", "You do not have permission to do this");
+      return res.redirect(`/campgrounds/${id}`);
+    }
     res.render("campgrounds/edit", { campground });
   })
 );
@@ -73,7 +84,12 @@ router.put(
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("error", "You do not have permission to do this");
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(
       id,
       { ...req.body.campground },
       { new: true }
@@ -87,8 +103,15 @@ router.delete(
   "/:id",
   isLoggedIn,
   catchAsync(async (req, res) => {
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("error", "You do not have permission to do this");
+      return res.redirect(`/campgrounds/${id}`);
+    }
+
     const { id } = req.params;
     await Campground.findByIdAndDelete(id);
+
     req.flash("success", "Successfully deleted campground");
     res.redirect(`/campgrounds`);
   })
